+++ R/format_table_glue.R
@@ -1,0 +1,299 @@
+# this function does the main composition of columns for the output.
+
+.format_glue_table <- function(x,
+                               style,
+                               coef_column = NULL,
+                               p_stars = NULL,
+                               ...) {
+  # evaluate dots
+  dots <- list(...)
+  original_x <- x
+
+  # default format
+  if (!is.null(dots$format)) {
+    format <- dots$format
+  } else {
+    format <- "text"
+  }
+
+  # column name for "pasted" columns
+  if (!is.null(dots$new_column_name)) {
+    new_column_name <- dots$new_column_name
+  } else {
+    new_column_name <- NULL
+  }
+
+  if (identical(format, "html")) {
+    linesep <- "<br>"
+  } else {
+    linesep <- " "
+  }
+
+  # init significant stars
+  if ((!is.null(style) && style %in% c("se", "ci")) || is.null(p_stars)) {
+    x$p_stars <- ""
+  } else {
+    x$p_stars <- p_stars
+  }
+
+  # potential names for the coefficient column
+  coefficient_names <- unique(c(
+    easystats_columns("estimate"),
+    broom_columns("estimate")
+  ))
+
+  # find columns
+  if (is.null(coef_column) || !coef_column %in% colnames(x)) {
+    coef_column <- intersect(colnames(x), coefficient_names)[1]
+  }
+  ci_column <- colnames(x)[endsWith(colnames(x), " CI") | colnames(x) == "CI"]
+  stat_colum <- colnames(x)[colnames(x) %in% c("t", "z", "Chi2", "Statistic") | grepl("^(t\\(|Chi2\\()", colnames(x))]
+  # modelbased
+  focal_term_column <- c(
+    attributes(original_x)$focal_terms,
+    attributes(original_x)$trend,
+    attributes(original_x)$contrast
+  )
+
+  # make sure we have a glue-like syntax
+  style <- .convert_to_glue_syntax(style, linesep)
+
+  # "|" indicates cell split
+  style <- unlist(strsplit(style, split = "|", fixed = TRUE))
+
+  # define column names
+  if (length(style) == 1 && !is.null(new_column_name)) {
+    column_names <- new_column_name
+  } else {
+    column_names <- .style_pattern_to_name(style, coef_column)
+  }
+
+  # paste glue together
+  formatted_columns <- lapply(seq_along(style), function(i) {
+    .format_glue_output(x, coef_column, ci_column, style[i], format, column_names[i])
+  })
+  out <- do.call(cbind, formatted_columns)
+
+  # add new_column_name to column names; for single column layout per model, we just
+  # need the column name. If the layout contains more than one column per model,
+  # add new_column_name in parenthesis.
+  if (!is.null(new_column_name) && nzchar(new_column_name, keepNA = TRUE)) {
+    if (ncol(out) > 1) {
+      colnames(out) <- paste0(colnames(out), " (", new_column_name, ")")
+    } else {
+      colnames(out) <- new_column_name
+    }
+  }
+
+  # remove empty parenthesis
+  out[] <- lapply(out, function(i) {
+    # here we either have "<br>" or " " as line breaks, followed by empty "()"
+    i <- gsub("<br>()", "", i, fixed = TRUE)
+    i <- gsub(" ()", "", i, fixed = TRUE)
+    i <- gsub("<br>(, )", "", i, fixed = TRUE)
+    i <- gsub(" (, )", "", i, fixed = TRUE)
+    i[i == "()"] <- ""
+    i[i == "(, )"] <- ""
+    # remove other non-matched patterns
+    i <- gsub("{stars}", "", i, fixed = TRUE)
+    i <- gsub("{rhat}", "", i, fixed = TRUE)
+    i <- gsub("{ess}", "", i, fixed = TRUE)
+    i <- gsub("{pd}", "", i, fixed = TRUE)
+    i <- gsub("{rope}", "", i, fixed = TRUE)
+    i
+  })
+
+  # define p-alike columns
+  p_column <- unique(c(easystats_columns("p"), broom_columns("p")))
+  uncertainty_column <- unique(c(
+    easystats_columns("uncertainty"),
+    broom_columns("uncertainty"),
+    ci_column
+  ))
+
+  # bind glue-columns to original data, but remove former columns first
+  original_x[c(coefficient_names, uncertainty_column, stat_colum, p_column)] <- NULL
+
+  # reorder
+  original_x <- standardize_column_order(original_x)
+
+  # we want: first parameter, then glue-columns, then remaining columns
+  parameter_column <- intersect(
+    c(
+      easystats_columns("parameter"),
+      broom_columns("parameter"),
+      focal_term_column
+    ),
+    colnames(original_x)
+  )
+  non_parameter_column <- setdiff(colnames(original_x), parameter_column)
+
+  cbind(original_x[parameter_column], out, original_x[non_parameter_column])
+}
+
+
+.convert_to_glue_syntax <- function(style, linesep = NULL) {
+  # set default
+  if (is.null(linesep)) {
+    linesep <- " "
+  }
+
+  # default
+  if (is.null(style)) {
+    style <- paste0("{estimate}", linesep, "({ci})|{p}")
+
+    # style: estimate and CI, p-value in separate column (currently identical to "ci_p2")
+  } else if (style %in% c("minimal", "ci_p2")) {
+    style <- paste0("{estimate}", linesep, "({ci})|{p}")
+
+    # style: estimate and CI, no p
+  } else if (style == "ci") {
+    style <- paste0("{estimate}", linesep, "({ci})")
+
+    # style: estimate, p-stars and CI
+  } else if (style == "ci_p") {
+    style <- paste0("{estimate}{stars}", linesep, "({ci})")
+
+    # style: estimate and SE, no p
+  } else if (style == "se") {
+    style <- paste0("{estimate}", linesep, "({se})")
+
+    # style: estimate, p-stars and SE
+  } else if (style == "se_p") {
+    style <- paste0("{estimate}{stars}", linesep, "({se})")
+
+    # style: estimate and SE, p-value in separate column
+  } else if (style %in% c("short", "se_p2")) {
+    style <- paste0("{estimate}", linesep, "({se})|{p}")
+
+    # style: only estimate
+  } else if (style %in% c("est", "coef")) {
+    style <- "{estimate}"
+  }
+
+  # replace \n for now with default line-separators
+  gsub("\n", linesep, style, fixed = TRUE)
+}
+
+
+.format_glue_output <- function(x, coef_column, ci_column, style, format, column_names) {
+  # separate CI columns, for custom layout
+  ci <- ci_low <- ci_high <- NULL
+  if (!insight::is_empty_object(ci_column)) {
+    ci <- x[[ci_column[1]]]
+    ci_low <- insight::trim_ws(gsub("(\\(|\\[)(.*),(.*)(\\)|\\])", "\\2", ci))
+    ci_high <- insight::trim_ws(gsub("(\\(|\\[)(.*),(.*)(\\)|\\])", "\\3", ci))
+  }
+
+  # fix p-layout
+  if ("p" %in% colnames(x)) {
+    x[["p"]] <- insight::trim_ws(x[["p"]])
+    x[["p"]] <- gsub("< .", "<0.", x[["p"]], fixed = TRUE)
+  }
+
+  # handle aliases
+  style <- tolower(style)
+  style <- gsub("{coef}", "{estimate}", style, fixed = TRUE)
+  style <- gsub("{coefficient}", "{estimate}", style, fixed = TRUE)
+  style <- gsub("{std.error}", "{se}", style, fixed = TRUE)
+  style <- gsub("{standard error}", "{se}", style, fixed = TRUE)
+  style <- gsub("{pval}", "{p}", style, fixed = TRUE)
+  style <- gsub("{p.value}", "{p}", style, fixed = TRUE)
+  style <- gsub("{ci}", "{ci_low}, {ci_high}", style, fixed = TRUE)
+
+  # align columns width for text format
+  .align_values <- function(i) {
+    if (!is.null(i)) {
+      non_empty <- !is.na(i) & nzchar(i, keepNA = TRUE)
+      i[non_empty] <- format(insight::trim_ws(i[non_empty]), justify = "right")
+    }
+    i
+  }
+
+  # we put all elements (coefficient, SE, CI, p, ...) in one column.
+  # for text format, where columns are not center aligned, this can result in
+  # misaligned columns, which looks ugly. So we try to ensure that each element
+  # is formatted and justified to the same width
+  if (identical(format, "text") || is.null(format)) {
+    x[[coef_column]] <- .align_values(x[[coef_column]])
+    x$SE <- .align_values(x$SE)
+    x[["p"]] <- .align_values(x[["p"]])
+    x$p_stars <- .align_values(x$p_stars)
+    ci_low <- .align_values(ci_low)
+    ci_high <- .align_values(ci_high)
+    x$pd <- .align_values(x$pd)
+    x$Rhat <- .align_values(x$Rhat)
+    x$ESS <- .align_values(x$ESS)
+    x$ROPE_Percentage <- .align_values(x$ROPE_Percentage)
+  }
+
+  # create new string
+  table_row <- rep(style, times = nrow(x))
+  for (r in seq_along(table_row)) {
+    table_row[r] <- gsub("{estimate}", x[[coef_column]][r], table_row[r], fixed = TRUE)
+    if (!is.null(ci_low) && !is.null(ci_high)) {
+      table_row[r] <- gsub("{ci_low}", ci_low[r], table_row[r], fixed = TRUE)
+      table_row[r] <- gsub("{ci_high}", ci_high[r], table_row[r], fixed = TRUE)
+    }
+    if ("SE" %in% colnames(x)) {
+      table_row[r] <- gsub("{se}", x[["SE"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("p" %in% colnames(x)) {
+      table_row[r] <- gsub("{p}", x[["p"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("p_stars" %in% colnames(x)) {
+      table_row[r] <- gsub("{stars}", x[["p_stars"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("pd" %in% colnames(x)) {
+      table_row[r] <- gsub("{pd}", x[["pd"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("Rhat" %in% colnames(x)) {
+      table_row[r] <- gsub("{rhat}", x[["Rhat"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("ESS" %in% colnames(x)) {
+      table_row[r] <- gsub("{ess}", x[["ESS"]][r], table_row[r], fixed = TRUE)
+    }
+    if ("ROPE_Percentage" %in% colnames(x)) {
+      table_row[r] <- gsub("{rope}", x[["ROPE_Percentage"]][r], table_row[r], fixed = TRUE)
+    }
+  }
+
+  # some cleaning: columns w/o coefficient are empty
+  table_row[x[[coef_column]] == "" | is.na(x[[coef_column]])] <- "" # nolint
+
+  # fix some p-value stuff, e.g. if pattern is "p={p]}",
+  # we may have "p= <0.001", which we want to be "p<0.001"
+  table_row <- gsub("=<", "<", table_row, fixed = TRUE)
+  table_row <- gsub("= <", "<", table_row, fixed = TRUE)
+  table_row <- gsub("= ", "=", table_row, fixed = TRUE)
+
+  # final output
+  x <- data.frame(table_row)
+  colnames(x) <- column_names
+
+  # sanity check - does column exist?
+  non_existent <- vapply(x, function(i) all(i == style), logical(1))
+  x[non_existent] <- NULL
+
+  x
+}
+
+
+.style_pattern_to_name <- function(style, coef_column) {
+  if (is.null(coef_column)) {
+    coef_column <- "Estimate"
+  }
+  column_names <- tolower(style)
+  # completely remove these patterns
+  column_names <- gsub("{stars}", "", column_names, fixed = TRUE)
+  # remove curlys
+  column_names <- gsub("{", "", column_names, fixed = TRUE)
+  column_names <- gsub("}", "", column_names, fixed = TRUE)
+  # manual renaming
+  column_names <- gsub("\\Qrope\\E", "% in ROPE", column_names)
+  column_names <- gsub("(estimate|coefficient|coef)", coef_column, column_names)
+  column_names <- gsub("\\Qse\\E", "SE", column_names)
+  column_names <- gsub("\\<ci\\>", "CI", column_names)
+  column_names <- gsub("<br>", "", column_names, fixed = TRUE)
+  column_names
+}